completion: improve handling of DWIM mode for switch/checkout
Both git switch and git checkout can "Do What I Mean" by interpreting
a name of unique remote branch as "please create and switch to a local
branch named after and tracking this remote branch".

Completion supports this DWIM mode by allowing to complete these unique
remote branch names. The current logic for when we decide to enable this
DWIM mode is not consistent between git checkout and git switch.

In both cases, this logic is enabled by default. Then, if any of
'--track', '--no-track', or '--no-guess' are provided on the command
line, then we disable it. Additionally, this logic is disabled when
GIT_COMPLETION_CHECKOUT_NO_GUESS is set to 1. For _git_switch, we also
unconditionally enable this logic if "--guess" is provided on the
command line.

The rules for when to enable or disable DWIM have gotten quite
complicated, so its time to refactor these into a function,
__git_checkout_default_dwim_mode().

First, the default is to enable DWIM logic. Then, if '--no-track' is
provided or if the GIT_COMPLETION_CHECKOUT_NO_GUESS variable is set, we
change to default to disabled.

A new helper, __git_find_last_on_cmdline is introduced, similar to the
already existing __git_find_on_cmdline, but which operates in reverse,
finding the *last* matching word of the provided wordlist.

This is then used to find the last --guess or --no-guess. In this way,
if --guess was last specified (and thus git option parsing will think
--guess is enabled), we will unconditionally enable DWIM mode. If
--no-guess is provided last, then we will unconditionally disable DWIM
mode. Finally, if neither are specified, we use the default.

This new logic is more robust, as we will correctly identify superseded
options, and ensure that both _git_switch and _git_checkout enable DWIM
in similar ways.

Add several tests which demonstrate the new expected behavior. Note that
some of the git switch tests are marked as failures because the default
git switch completion with --guess is sub-par as discussed in a previous
commit. This will be fixed in a future change.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>