completion: improve handling of -c/-C and -b/-B in switch/checkout
Currently, completion for git switch and git checkout do not take into
account the branch creation options, -c/-C and -b/-B respectively.

The form of git switch with -c or -C is to take a branch name followed
by a starting point. For this reason, when completing -c or -C, we ought
to be able to complete the start point as any arbitrary reference.

However, we probably do not want to complete the actual argument for
these options with any reference, as this would provide too many
unexpected items.

Use the wordlist to determine the previous word on the command line. If
it's -c or -C (-b/-B for checkout), then we know that we are completing
the argument for the branch name. One possible assumption is that we
have no valid completion. But consider that in many cases a user may
wish to complete a branch by editing the name of a pre-existing branch.

Given that a user who already knows the branch name they want to
complete will simply not use completion, it makes sense to complete the
small subset of local branches when completing the argument for -c/-C.

In all other cases, if -c/-C are on the command line but are not the
most recent option, then we must be completing a start-point, and should
allow completing against all references.

Add many tests to cover various behavior and interaction around enabling
or disabling DWIM mode.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>